Prefer local image when present
Resolves #177

This PR changes the behavior of `copa patch`. When an image is available
in the local *docker* image store, that image will be preferred to one
in the remote registry.

Importantly, this is only happening for the local docker store. If the
image is some other store, for example `containerd`, it will not be
preferred to the remote version.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>