Add integration tests
This commit adds the `"localName"` field to the integration test spec.
When the field is present, the image is pulled and tagged with the value
of the field. Then the scans and copa invocation use the value instead
of the supplied registry, repo, and tag.

The following three names are supplied as tests:
`registry.copacetic.test/repo/image:tag`
`local/image:tag`
`localimage:tag`

Since `registry.copacetic.test` cannot be a valid domain name due to the
`.test` suffix, the image will never resolve remotely. Without domains,
neither of the other two references will resolve remotely either.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>